added new modules - VPN Select List

+++ mobile-client-android/sentinel_android/app/src/main/java/sentinelgroup/io/sentinel/network/model/Location.java
@@ -1,8 +1,8 @@
 package sentinelgroup.io.sentinel.network.model;
 
 public class Location {
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
     public String city;
     public String country;
 }
