+++ mobile-client-android/sentinel_android/app/src/main/java/sentinelgroup/io/sentinel/network/model/NetSpeed.java
@@ -1,6 +1,6 @@
 package sentinelgroup.io.sentinel.network.model;
 
 public class NetSpeed {
-    public float download;
-    public float upload;
+    public double download;
+    public double upload;
 }
